role group

+++ app/Http/Controllers/GroupController.php
@@ -9,8 +9,6 @@ use App\Http\Requests;
 use App\Http\Controllers\Controller;
 use App\Acl\Eloquent\Group;
 
-use Sentinel;
-
 class GroupController extends Controller
 {
     /**
@@ -64,10 +62,6 @@ class GroupController extends Controller
     public function show($id)
     {
         $group = Group::find($id);
-        //if (!$role || $project_key != $role->project_key)
-        //{
-        //    throw new \UnexpectedValueException('the role does not exist or is not in the project.', -10002);
-        //}
         return Response()->json([ 'ecode' => 0, 'data' => $group ]);
     }
 
@@ -104,7 +98,14 @@ class GroupController extends Controller
      */
     public function destroy($id)
     {
+        $group = Group::find($id);
+        if (!$group)
+        {
+            throw new \UnexpectedValueException('the group does not exist.', -10002);
+        }
+
         Group::destroy($id);
+        Event::fire(new DelUserEvent($id));
         return Response()->json([ 'ecode' => 0, 'data' => [ 'id' => $id ] ]);
     }
